requirements.txt
rdflib~=7.1.3
geomet~=1.1.0
Unidecode~=1.3.8
requests~=2.32.3
click~=8.1.8
molgenis-emx2-pyclient>=11.57.0
python-dotenv~=1.0.1